accounts/urls.py
from django.contrib.auth.views import LogoutView
from django.urls import path
from . import views

urlpatterns = [
    path('register/', views.register, name='register'),
    path('login/', views.user_login, name='login'),
    path('logout/', LogoutView.as_view(), name='logout'),
    path('profile/', views.profile, name='profile'),
    path('export/', views.export_data, name='export_data'),
    path('supprimer/', views.request_deletion, name='request_deletion'),
]
